Handle postgres arrays correctly.
The current behavior for postgres.ARRAY sort of works. SQLAlchemy knows
that the `python_type` of `ARRAY` columns is `list`, so we convert them
to `List` fields. But we currently ignore the nested types:
`ARRAY(Integer)` should convert to `List(Int)`, and `ARRAY(String)`
should convert to `List(Str)`. This patch inspects the nested type of
the `ARRAY` to build the appropriate `List`.