Add interface to update existing rows.
Currently, primary keys are set to dump but not load, such that
deserializing always creates a new object. This patch changes the
default behavior, such that we attempt to load by primary key on
deserialization, and update the returned object if found. This default
can also be overridden by passing `instance` to either a `ModelSchema`
constructor, or to its `load` method; updates will then be applied to
`instance`, and deserializing will not attempt to load records from the
database.