Require non-nullable columns without default values.
Currently, we require values for all non-nullable columns to be set.
However, as pointed out by @AdrielVelazquez, this leads to incorrect
behavior for non-nullable columns with default values. This patch sets
columns as required only when non-nullable and no default value is set,
either using the `default` or `server_default` settings, or if the
column uses `auto_increment`.

[Resolves #47]