Fix example for SQLAlchemySchema (#283)
* Fix example for SQLAlchemySchema

Hi, I just noticed what appears to be an issue in the documented examples in the new `sqlalchemy_schema` modules. 

The example for `SQLAlchemySchema` imports `SQLAlchemyAutoSchema` and subclasses `SQLAlchemy` in the `UserSchema` class. Also, in that example and the `SQLAlchemyAutoSchema` example, I removed the `session` import as neither example makes use of it. I could add the `sqla_session` option to the `class Meta` in each example if you'd rather have it there?

Thanks for the helpful lib!

* Add @5uper5hoot to AUTHORS

Co-authored-by: Steven Loria <sloria1@gmail.com>